requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
redis